스타일 분리 완
