+++ 3-weekly-mission/.env
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_KAKAO_JS_AppKey= ff1f8f88c3baf99238f45646d198c927
