+++ 3-weekly-mission/src/components/footer/LinkLogo/LinkLogo.tsx
