+++ 3-weekly-mission/src/components/footer/LinkString/LinkString.module.css
@@ -1,0 +1,11 @@
+.footer-links {
+  display: flex;
+  column-gap: 3rem;
+  padding-right: 1.8rem;
+}
+
+.footer-link {
+  color: #cfcfcf;
+  font-family: Arial;
+  font-size: 1.6rem;
+}
