+++ 3-weekly-mission/src/components/footer/Sns/Sns.module.css
@@ -1,0 +1,5 @@
+.sns {
+  display: flex;
+  column-gap: 1.2rem;
+  height: 2rem;
+}
