+++ 3-weekly-mission/src/components/header/Folder/Folder.module.css
@@ -1,0 +1,65 @@
+.information {
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  gap: 0.8rem;
+  padding: 11.4rem 0 6rem;
+
+  background: var(--linkbrary-bg);
+}
+
+.folder-info {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 2rem;
+}
+
+.user-info {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1.2rem;
+}
+
+.owner-profile {
+  width: 6rem;
+  height: 6rem;
+  position: relative;
+}
+
+.owner-name {
+  color: var(--text-color-light-mode, #000);
+  font-size: 1.6rem;
+  line-height: 2.4rem; /* 150% */
+}
+
+.folder-name {
+  color: #000;
+  text-align: center;
+  font-feature-settings: 'clig' off, 'liga' off;
+  font-size: 4rem;
+  font-weight: 600;
+}
+
+@media (max-width: 767px) {
+  .owner-profile {
+    width: 4rem;
+    height: 4rem;
+  }
+
+  .owner-name {
+    color: var(--text-color-light-mode, #000);
+    font-size: 1.4rem;
+  }
+
+  #folder-name {
+    color: #000;
+    text-align: center;
+    font-size: 3.2rem;
+    font-weight: 600;
+    letter-spacing: -0.2px;
+  }
+}
