+++ 3-weekly-mission/src/components/header/Nav/Nav.module.css
@@ -1,0 +1,106 @@
+.nav {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  max-width: 192rem;
+  margin: 0 auto;
+  height: 9.4rem;
+  padding: 0 20rem;
+  position: static;
+  top: 0;
+  left: 0;
+  right: 0;
+  z-index: 1;
+}
+
+.nav:not(.not-fixed) {
+  position: fixed;
+}
+
+.nav.not-fixed {
+  background: var(--linkbrary-bg);
+}
+
+.nav .logo {
+  width: 13.3rem;
+  height: 2.4rem;
+  position: relative;
+}
+
+.cta {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 5.4rem;
+  cursor: pointer;
+  background-image: linear-gradient(135deg, #6d6afe 0%, #6ae3fe 100%);
+  border-radius: 0.8rem;
+  color: #f5f5f5;
+  font-size: 1.8rem;
+  font-weight: 600;
+}
+
+.cta-short {
+  width: 12.8rem;
+}
+
+@media (max-width: 767px) {
+  .nav {
+    height: 6.3rem;
+    padding: 0 3.2rem;
+  }
+
+  .nav .logo {
+    width: 8.8667rem;
+    height: 1.6rem;
+  }
+
+  .user .email {
+    display: none;
+  }
+
+  .cta {
+    height: 3.7rem;
+    font-size: 1.4rem;
+  }
+
+  .cta-short {
+    width: 8rem;
+  }
+
+  .footer-box {
+    display: grid;
+    grid-template-areas:
+      'links sns'
+      'copyright .';
+    row-gap: 6rem;
+    padding: 0 3.2rem;
+  }
+
+  .copyright {
+    grid-area: copyright;
+    font-weight: 400;
+  }
+
+  .footer-links {
+    grid-area: links;
+  }
+
+  .footer-link {
+    font-size: 1.6rem;
+    font-weight: 400;
+  }
+}
+
+/* nav바 반응형 여백 */
+@media (max-width: 1199px) {
+  .nav {
+    padding: 0 calc((100% - 78.32rem) / 2);
+  }
+}
+@media (max-width: 864px) {
+  .nav {
+    padding: 0 3.2rem;
+  }
+}
