+++ 3-weekly-mission/src/components/input/input.tsx
@@ -71,4 +71,4 @@ export default function Input({ type }: Props) {
       )}
     </div>
   );
-}s
+}
